Refactor and add test files
